IT WORKS NOW!!!!! - moved ssbo copy to a new function in montecarlo, it has to be called before the first compute call is dispatched

+++ src/monteCarlo/MonteCarloParams.hpp
@@ -3,7 +3,7 @@
 
 struct MonteCarloParameters {
 	int vertexN;
-	int eps;
+	float eps;
 	int sampleN;
 	int maxWalk;
 };
